Added queue back in.
NOTE: There might be an issue here where scalars are not valid when
valid data is captured. This seems to only occur when the quee is added.
However switching to non-blocking input seems to have partially rectified this
issue since the scalars are ready by the time valid data is seen.